Fix ApiJob cleanup logic. ApiJob doesn't have a status field anymore, so modify_date won't get updated. Put modify_date on ApiJobUnit instead.

+++ project/api_core/tasks.py
@@ -13,8 +13,19 @@ from .models import ApiJob
     ignore_result=True,
 )
 def clean_up_old_api_jobs():
+    """
+    Clean up API jobs that satisfy both of these criteria:
+    1. Job was created 30+ days ago
+    2. All of the job's job units were modified 30+ days ago
+
+    Note that 2 is not sufficient in the corner case where a job has been
+    created, but its job units haven't been created yet.
+    """
     thirty_days_ago = timezone.now() - timedelta(days=30)
-    for job in ApiJob.objects.filter(modify_date__lt=thirty_days_ago):
-        # Job was last modified over 30 days ago.
-        # Delete the job, and its job units should cascade-delete with it.
-        job.delete()
+
+    for job in ApiJob.objects.filter(create_date__lt=thirty_days_ago):
+        units_were_modified_in_last_30_days = job.apijobunit_set.filter(
+            modify_date__gt=thirty_days_ago).exists()
+        if not units_were_modified_in_last_30_days:
+            # Delete the job, and its job units should cascade-delete with it.
+            job.delete()
